command: append object error to multierror to return non-zero exit code error (#328)
Creating two error objects instead of one, is to avoid data race. Since there is a goroutine running which tries to write merror object, there might be a data race for merror object in cp, rm and select commands. We can create 1 error object for the goroutine and 1 for the main routine, and append them at the end of the execution.

Tested #304 's first case, it returns exit code 1.

Resolves #304