add "--all-versions" flag to ls, rm and du subcommands to support versioning
This commit adds "--all-versions" flag to "ls", "rm", "du" subcommands.

When rm is used with "--all-versions" flag, it succesfully deletes all "versions" of object, but it DOES NOT delete any of the "Delete Markers" associated with the given key(s). For now, I'm not sure about how to handle delete markers. Whether deletion of them should be a default behaviour, or should we require another flag for them or maybe another subcommand.

Similarly, we are yet to decide how or whether s5cmd will support enabling or suspending versioning.

As a last note, for operations with "--all-versions" I'm using same interfaces/methods used by default. So I expect them to have similar high performance.

Updates. #386, #218.