storage/url: fixes the bug where some part of destination path is removed by cp (#456)
The cause of problem is a little bit weird.

(url).SetRelative method was given the source (abs.) path as it was provided in command line arguments including the wildcard characters(* and ?). Hence it determined relative path incorrectly, where "expected" relative path became "../expected" which in turn caused the some part of destination to be lost (since path/filepath.Join converts e.g. /a/b/../c to /a/c)

We also need to identify whether the * and ? characters present in base strings are part of key or they are used for wildcard operation. So we need to know if the "raw" flag is given. To. solve this I changed (URL).SetRelative method to take argument of type URL rather than string.

Fixes #360

Co-authored-by: Selman Kayrancioglu <seruman@users.noreply.github.com>
Co-authored-by: Aykut Farsak <aykutfarsak@gmail.com>
Co-authored-by: Muhammed Can Küçükaslan <32893303+MuhammedCanKucukaslan@users.noreply.github.com>