fix the bug caused by unconsumed channel
The destination only objects are only needed if they're going to be deleted.  So the consumer of onlyDest channel worked if they're going to be deleted.
This naive behaviour caused channel not to accept URLs and block send requests when the channel buffer is filled!

Moral of the story, don't be naive. The fact that you need channel in one case does not mean you need to consume channel only in that case.