use custom method to capitalize first letter
Our use case is very simple then the the problem that text/cases is handling. We only want first rune to be upper case, and rest to be lower case, we don't care the word breaks and we don't care any other language than English, in fact we only care words: "Suspended" and "Enabled".
So it would be overkill to use that package.