feat: dependabot workflow automation for updating dependency (#530)
Manual control of dependency is fine but with a growing no. of distributed upstream dependencies, it is hard to manage. So, for that automation should be there to update dependencies. Further, CI Pipeline is there to test those changes. The current build is having many old dependencies due to that many vulnerabilities were found. i.e. why workflow automation will help here.
Further, this will not update the dependencies automatically, instead, a PR will be opened with changes that can be reviewed and tested with CI.

Signed-off-by: Pratik Raj rajpratik71@gmail.com