fix: sync glacier objects overwrite (#734)
This pull request addresses a bug (#712) in the sync command where
objects in Glacier storage at the destination were being overwritten
during synchronization. closes #712

## Bug Details:

During the sync process, if an object exists in both the source and
destination, and the destination object is in Glacier storage, it was
incorrectly being overwritten by the source object.

## Fixes:

- Modified the function that checks which objects to compare. 

Previously, if either the source or destination object was in Glacier
storage, it was skipped. In the new version, only the source object is
skipped, but the destination object is not. If the destination object is
skipped, it will be considered as only in the source during the
comparison phase, leading to it being overwritten by the source object.

- Added tests for this specific case to ensure correctness.

---------

Co-authored-by: İbrahim Güngör <igungor@gmail.com>
Co-authored-by: İlkin Balkanay <ilkinulas@gmail.com>
Co-authored-by: S.Burak Yaşar <burakyasar@peak.com>